Introduce Highlighter class structure and organize files under highlighters directory. Fix Python regex issues for multi-line comments.
