+++ CustomTextEditor/highlighters/pythonhighlighter.h
